+++ src/features/advanced/index.ts
@@ -23,4 +23,5 @@ import './lm-hide-rating';
 import './mat-clean-info';
 import './minimize-headers/minimize-headers';
 import './nots-clean-notifications';
+import './prod-hide-percent';
 import './shpf-hide-sort-options';